+++ src/components/ContactFormStyles.css
@@ -23,17 +23,6 @@
   border: 1px solid #2a2a2a;
 }
 
-.conatct button {
-  text-decoration: none;
-  background-color: lightgray;
-  padding: 1rem 5rem;
-  border-radius: 6px;
-  font-size: 1.1rem;
-  font-weight: bold;
-  letter-spacing: 2px;
-  color: black;
-  margin-bottom: 3rem;
-}
 .conatct button:hover {
   background-color: #01959a;
   color: #fff;
