+++ src/pages/About.js
@@ -2,7 +2,7 @@
 // import Hero from "../components/Hero";
 // import Navbar from "../components/Navbar";
 // import Footer from "../components/Footer";
-// import img1 from "../asserts/12.jpg";
+// import img1 from "../assets/images//12.jpg";
 // import Aboutus from "../components/Aboutus";
 
 // const About = () => {
@@ -21,8 +21,8 @@ import React from "react";
 import Hero from "../components/Hero";
 import Navbar from "../components/Navbar";
 import Footer from "../components/Footer";
-import img1 from "../asserts/12.jpg";
-import img2 from "../asserts/9.jpg";
+import img1 from "../assets/images//12.jpg";
+import img2 from "../assets/images//9.jpg";
 import Aboutus from "../components/Aboutus";
 
 const About = () => {
